Face Recognition Flask App/requirements.txt
Flask==2.2.3
face_recognition==1.3.0
opencv-python==4.7.0.72
numpy==1.23.5
Pillow==9.3.0
werkzeug== 2.2.3